cfs/file_slicer/requirements.txt
absl-py
google-cloud-storage
